+++ src/_modules/root/root.controller.ts
@@ -4,13 +4,16 @@ import consts from '~config/consts';
 class RootController {
   constructor(){}
 
-  async helloWorld({set, user, session}: any) {
+  async helloWorld({set, user, session, request:{ headers }}: any) {
     try {
       console.log("User: ",user);
       console.log("Session: ",session);
       console.log("ProfileID: ",user?.profileId ?? null);
+      console.log(headers.get('accept'));
       
-      return { message: `Welcome to the ${consts.server.name} Server! Version ${consts.server.version}` }
+      const isBrowser = headers.get('accept').includes('text/html');
+      
+      return isBrowser ? Bun.file('public/welcome.html') : { message: `Welcome to the ${consts.server.name} Server! Version ${consts.server.version}` }
     } catch (error) {
       console.error(error);
 
@@ -41,6 +44,29 @@ class RootController {
 
     try {
       return { message: `It's a beautiful ${getGreeting()} in ${timezone ?? 'your area'} isn't it ${params?.name || ''}?` }
+    } catch (error) {
+      console.error(error);
+
+      return { message: `Error occurred` }
+    }
+  }
+
+  async htmx({ hx, ip }: any) {
+    
+    console.log(ip);
+    console.log(hx);
+    try {
+      return { data: hx ? {
+        request: hx.request,
+        boosted: hx.boosted,
+        historyRestoreRequest: hx.historyRestoreRequest,
+        currentURL: hx.currentURL,
+        prompt: hx.prompt,
+        target: hx.target,
+        triggerName: hx.triggerName,
+        trigger: hx.trigger,
+        isHTMX: hx.isHTMX,
+      } : 'No HTMX data', message: `HTMX test route. IP: ${ip ?? 'Unknown'}` }
     } catch (error) {
       console.error(error);