+++ src/_modules/root/root.handler.ts
@@ -1,23 +1,27 @@
 
 import Elysia, { t } from "elysia";
 import { RootController } from "~modules/root";
-import { checkAuth, checkForProfile, checkJWTAuth } from "~middleware/authChecks";
+import { checkAuth, checkForProfile } from "~middleware/authChecks";
 
-const rootHandler = new Elysia({ prefix: '' })
+const rootHandler = new Elysia({ prefix: '', tags: ['root'] })
 
     .get('/', RootController.helloWorld, {
-        detail: {description: 'Hello World', tags: ['root'] },
+        detail: {description: 'Hello World' },
     }) // main route
 
     .get('/hello', RootController.helloTime, {
         beforeHandle: [checkAuth],
-        detail: {description: 'Greeting + time', tags: ['root'] },
+        detail: {description: 'Greeting + time' },
     }) // hello route
       
     .get('/hello/:name', RootController.helloTime, {
         beforeHandle: [checkAuth, checkForProfile],
-        detail: {description: 'Greeting + time', tags: ['root'] },
+        detail: {description: 'Greeting + time' },
         params: t.Object({ name: t.String() })
     }) // hello + name route
+    
+    .get('/htmx', RootController.htmx, {
+        detail: {description: 'HTMX test route' }
+    }) // htmx test route
 
 export default rootHandler;